% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMRASPackage.R
\docType{data}
\name{IMRAS_counts_Eset}
\alias{IMRAS_counts_Eset}
\title{Non-normalized and unfiltered raw read counts of IMRAS  study.}
\format{a \code{ExpressionSet} containing the following fields:
\describe{assayData and phenoData}}
\source{
The data comes from RNASeq fasta files.
}
\description{
Contains two tables: assayData which contains raw read counts where the
rows are genes and columns are samples and phenoData which contains
experimental design metadata where each row is a sample and each column
is a factor describing each sample. The rows of phenoData map to the
columns of assayData. The assayData and phenoData tables can be extracted
with \code{exprs(IMRAS_counts_Eset)} and \code{pData(IMRAS_counts_Eset)} respectively.
}
\seealso{
\link{IMRASPackage}
\link{IMRAS_tpm_Eset}
}
