% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMRASPackage.R
\name{IMRASPackage}
\alias{IMRASPackage}
\title{IMRASPackage}
\description{
Raw read counts and TPM values for RNASeq IMRAS  study.
}
\details{
IMRAS Package
A data package for IMRAS Package.

Use \code{data(package='IMRAS Package')$results[, 3]} to see a
list of availabledata sets in this data package and/or
DataPackageR::load_all _datasets() to load them. Alternatively one can use
\code{data(IMRAS_counts_Eset)} or \code{data(IMRAS_tpm_Eset)} to load the data.
}
\seealso{
\link{IMRAS_counts_Eset}
\link{IMRAS_tpm_Eset}
}
